[lib/base] Add typecasting to size_t before applying OT_UNSIGNED

+++ sw/device/lib/base/memory.c
@@ -20,8 +20,8 @@ static size_t compute_num_leading_bytes(const void *left, const void *right,
   if (len < alignof(uint32_t)) {
     return len;
   }
-  const size_t left_ahead = OT_UNSIGNED(misalignment32_of((uintptr_t)left));
-  const size_t right_ahead = OT_UNSIGNED(misalignment32_of((uintptr_t)right));
+  const size_t left_ahead = OT_UNSIGNED((size_t)misalignment32_of((uintptr_t)left));
+  const size_t right_ahead = OT_UNSIGNED((size_t)misalignment32_of((uintptr_t)right));
   if (right == NULL || left_ahead == right_ahead) {
     return (4 - left_ahead) & 0x3;
   }
